requirements.txt
# Deep Learning Urban Change Detection Project
# Python 3.8+ recommended

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Jupyter (optional, for experimentation)
jupyter>=1.0.0
ipywidgets>=8.1.0

